+++ tests/environments/storage/test_docker_storage.py
@@ -111,6 +111,14 @@ def test_initialized_docker_storage():
     assert storage.local_image
 
 
+def test_docker_storage_allows_for_user_provided_config_locations():
+    storage = Docker(env_vars={"PREFECT__USER_CONFIG_PATH": "1"},)
+
+    assert storage.env_vars == {
+        "PREFECT__USER_CONFIG_PATH": "1",
+    }
+
+
 def test_files_not_absolute_path():
     with pytest.raises(ValueError):
         storage = Docker(files={"test": "test"})
@@ -316,7 +324,7 @@ def test_create_dockerfile_from_base_image():
             "master",
             (
                 "FROM python:3.6-slim",
-                "pip install git+https://github.com/PrefectHQ/prefect.git@master",
+                "pip show prefect || pip install git+https://github.com/PrefectHQ/prefect.git@master",
             ),
         ),
         (
@@ -324,7 +332,7 @@ def test_create_dockerfile_from_base_image():
             (
                 "FROM python:3.6-slim",
                 "apt update && apt install -y gcc git && rm -rf /var/lib/apt/lists/*",
-                "pip install git+https://github.com/PrefectHQ/prefect.git@424be6b5ed8d3be85064de4b95b5c3d7cb665510#egg=prefect[kubernetes]",
+                "pip show prefect || pip install git+https://github.com/PrefectHQ/prefect.git@424be6b5ed8d3be85064de4b95b5c3d7cb665510#egg=prefect[kubernetes]",
             ),
         ),
     ],